add comments, move mappings to store
